OCPBUGS-79363: Expand tree shaking unit test
To include checking that `monaco-editor`, `xterm`, and `@octokit/rest` (part of git-service) aren't included alongside every page load
